chore: run CI on windows
