+++ .github/workflows/verify.yml
@@ -3,8 +3,8 @@ name: Verify changes
 on: pull_request
 
 jobs:
-  verify:
-    name: Verify changes
+  verify-linux:
+    name: Verify linux
     runs-on: ubuntu-latest
     strategy:
       matrix:
@@ -32,6 +32,9 @@ jobs:
       - name: Install dependencies
         run: yarn --frozen-lockfile
 
+      - name: Build packages
+        run: yarn build
+
       # build for production in CI to make sure tests can run with production build
       - name: Build for production
         run: yarn build:production
@@ -41,6 +44,39 @@ jobs:
 
       - name: Test
         run: yarn test
+
+  verify-windows:
+    name: Verify windows
+    runs-on: windows-latest
+    steps:
+      - uses: actions/checkout@v2
+
+      - name: Setup Node 12.x
+        uses: actions/setup-node@v1
+        with:
+          node-version: 12.x
+
+      - name: Get yarn cache directory path
+        id: yarn-cache-dir-path
+        run: echo "::set-output name=dir::$(yarn cache dir)"
+
+      - uses: actions/cache@v2
+        id: yarn-cache # use this to check for `cache-hit` (`steps.yarn-cache.outputs.cache-hit != 'true'`)
+        with:
+          path: ${{ steps.yarn-cache-dir-path.outputs.dir }}
+          key: ${{ runner.os }}-yarn-${{ hashFiles('**/yarn.lock') }}
+          restore-keys: |
+            ${{ runner.os }}-yarn-
+
+      - name: Install dependencies
+        run: yarn --frozen-lockfile
+
+      - name: Build
+        run: yarn build
+
+      - name: Test
+        run: yarn test:windows
+
   verify-browserstack:
     name: Verify browserstack
     runs-on: ubuntu-latest
@@ -66,6 +102,9 @@ jobs:
 
       - name: Install dependencies
         run: yarn --frozen-lockfile
+
+      - name: Build packages
+        run: yarn build
 
       - name: Test
         run: |
