+++ package.json
@@ -4,7 +4,7 @@
   "license": "MIT",
   "scripts": {
     "build": "rimraf packages/*/tsconfig.tsbuildinfo && tsc --build",
-    "build:production": "node scripts/workspaces-scripts.mjs build:production",
+    "build:production": "node scripts/workspaces-scripts-bin.mjs build:production",
     "build:site": "rocket-build",
     "build:watch": "rimraf packages/*/tsconfig.tsbuildinfo && tsc --build --watch",
     "format": "npm run format:eslint && npm run format:prettier",
@@ -14,12 +14,13 @@
     "lint": "npm run lint:eslint && npm run lint:prettier",
     "lint:eslint": "eslint --ext .ts,.js .",
     "lint:prettier": "node node_modules/prettier/bin-prettier.js \"**/*.{ts,js,md}\" --check --ignore-path .eslintignore",
-    "postinstall": "yarn install-puppeteer-firefox && yarn build && patch-package",
+    "postinstall": "patch-package",
     "reinstall-workspace": "rimraf packages/*/node_modules && rimraf node_modules && yarn install && yarn build",
     "release": "changeset publish && yarn format",
     "start": "rocket-start",
-    "start:build":"es-dev-server --root-dir _site --compatibility none --open",
-    "test": "node scripts/workspaces-scripts.mjs test:ci",
+    "start:build": "es-dev-server --root-dir _site --compatibility none --open",
+    "test": "node scripts/workspaces-scripts-bin.mjs test:ci",
+    "test:windows": "node scripts/runWindowsTests.mjs",
     "update-dependency": "node scripts/update-dependency.js",
     "update-package-configs": "ts-node scripts/update-package-configs.ts"
   },
