+++ packages/dev-server-esbuild/test/ts.test.ts
@@ -7,6 +7,8 @@ import { expectIncludes, expectNotIncludes } from '@web/dev-server-core/src/test
 import { esbuildPlugin } from '../src/esbuildPlugin';
 
 describe('esbuildPlugin TS', function () {
+  this.timeout(5000);
+
   it('transforms .ts files', async () => {
     const { server, host } = await createTestServer({
       rootDir: __dirname,
