+++ packages/test-runner-commands/test/emulate-media/emulateMediaPlugin.test.ts
@@ -1,4 +1,5 @@
 import path from 'path';
+import { platform } from 'os';
 import { runTests } from '@web/test-runner-core/dist/test-helpers';
 import { chromeLauncher } from '@web/test-runner-chrome';
 import { playwrightLauncher } from '@web/test-runner-playwright';
@@ -6,7 +7,7 @@ import { testRunnerServer } from '@web/test-runner-server';
 
 import { emulateMediaPlugin } from '../../src/emulateMediaPlugin';
 
-describe.only('setViewportPlugin', function test() {
+describe('emulateMediaPlugin', function test() {
   this.timeout(20000);
 
   it('can emulate media on puppeteer', async () => {
@@ -21,20 +22,23 @@ describe.only('setViewportPlugin', function test() {
     );
   });
 
-  it('can emulate media on playwright', async () => {
-    await runTests(
-      {
-        browsers: [
-          playwrightLauncher({ product: 'chromium' }),
-          playwrightLauncher({ product: 'firefox' }),
-          // TODO: make webkit work in the CI
-          // playwrightLauncher({ product: 'webkit' }),
-        ],
-        server: testRunnerServer({
-          plugins: [emulateMediaPlugin()],
-        }),
-      },
-      [path.join(__dirname, 'browser-test.js')],
-    );
-  });
+  // playwright doesn't work on windows VM right now
+  if (platform() !== 'win32') {
+    it('can emulate media on playwright', async () => {
+      await runTests(
+        {
+          browsers: [
+            playwrightLauncher({ product: 'chromium' }),
+            playwrightLauncher({ product: 'firefox' }),
+            // TODO: make webkit work in the CI
+            // playwrightLauncher({ product: 'webkit' }),
+          ],
+          server: testRunnerServer({
+            plugins: [emulateMediaPlugin()],
+          }),
+        },
+        [path.join(__dirname, 'browser-test.js')],
+      );
+    });
+  }
 });
