+++ packages/test-runner-commands/test/set-viewport/setViewportPlugin.test.ts
@@ -1,4 +1,5 @@
 import path from 'path';
+import { platform } from 'os';
 import { runTests } from '@web/test-runner-core/dist/test-helpers';
 import { chromeLauncher } from '@web/test-runner-chrome';
 import { playwrightLauncher } from '@web/test-runner-playwright';
@@ -21,20 +22,23 @@ describe('setViewportPlugin', function test() {
     );
   });
 
-  it('can set the viewport on playwright', async () => {
-    await runTests(
-      {
-        browsers: [
-          playwrightLauncher({ product: 'chromium' }),
-          playwrightLauncher({ product: 'firefox' }),
-          // TODO: make webkit work in the CI
-          // playwrightLauncher({ product: 'webkit' }),
-        ],
-        server: testRunnerServer({
-          plugins: [setViewportPlugin()],
-        }),
-      },
-      [path.join(__dirname, 'browser-test.js')],
-    );
-  });
+  // playwright doesn't work on windows VM right now
+  if (platform() !== 'win32') {
+    it('can set the viewport on playwright', async () => {
+      await runTests(
+        {
+          browsers: [
+            playwrightLauncher({ product: 'chromium' }),
+            playwrightLauncher({ product: 'firefox' }),
+            // TODO: make webkit work in the CI
+            // playwrightLauncher({ product: 'webkit' }),
+          ],
+          server: testRunnerServer({
+            plugins: [setViewportPlugin()],
+          }),
+        },
+        [path.join(__dirname, 'browser-test.js')],
+      );
+    });
+  }
 });
