+++ scripts/runWindowsTests.mjs
@@ -1,0 +1,11 @@
+import { runWorkspacesScripts } from './runWorkspacesScripts.mjs';
+
+runWorkspacesScripts({
+  script: 'test:ci',
+  concurrency: 1,
+  filteredPackages: [
+    // playwright can't install browsers on my windows VM
+    // 'test-runner-playwright',
+    // 'test-runner-selenium',
+  ],
+});
