+++ scripts/workspaces-scripts-bin.mjs
@@ -1,0 +1,4 @@
+import { runWorkspacesScripts } from './runWorkspacesScripts.mjs';
+
+const script = process.argv[process.argv.length - 1];
+runWorkspacesScripts({ script, concurrency: 5 });
